Extract client code
For the next test we were going to write, we realised we needed to think about a creating a client-side API and associated functionality.
In doing so, we realised that the cucumber test actually had client-side code in it, so our first step was to extract that, making the test code much cleaner.

Also removed a superfluous function that wasn't being called, and changed "var" to "let" to limit variable scope.