Start local environment automation
The bash scripts are flaky and hard to maintain.

Automating the test environment setup using dockerode and Node is the alternative path we're trying. The local-environment.js script currently stops and removes the DynamoDB container, and removes the Docker local network successfully.

Need to refactor for cleanliness next, and also ensure the Docker local network is removed if there is no existing DynamoDB container.