Add initial test to establish websocket and attempt login
When the client attempts to connect, if the user is not registered, the websocket returns a 403 "Forbidden" error and closes the socket.

The test does the simplest thing for the moment, which is return the 403 (no check if the user is actually registered)

Also separated the code into "registration" and "chat" modules to indicate the different concerns. Logging the user in is part of the "chat" module.